Start cleaning up the structure. Still need to organise things into classes.
