+++ include/formats.h
@@ -1,0 +1,14 @@
+struct FilmFormat
+{
+  int id;
+  String name;
+  int sensor[22];
+  int frame[22];
+};
+
+FilmFormat film_formats[] = {
+    {35, "PANO", {0, 35, 65, 95, 125, 155, 185, 215, 245, 275, 305, 335, 365, 395, 425, 455, 485, 515, 545, 575, 605, 700}, {0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 99}},
+    {645, "6x4.5", {0, 130, 157, 183, 208, 232, 255, 277, 298, 318, 337, 355, 372, 388, 403, 550}, {0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 99}},
+    {66, "6x6", {0, 135, 165, 194, 222, 249, 275, 300, 324, 347, 367, 386, 404, 550}, {0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 99}},
+    {67, "6x7", {0, 137, 171, 204, 236, 267, 297, 326, 354, 381, 407, 550}, {0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 99}},
+};
