+++ include/hardware.h
@@ -1,0 +1,23 @@
+// Hardware init
+// ---------------------
+// Inputs
+ESP32Encoder encoder;
+Bounce2::Button lbutton = Bounce2::Button();
+Bounce2::Button rbutton = Bounce2::Button();
+
+// Battery gauge
+Adafruit_MAX17048 maxlipo;
+
+// Lightmeter
+BH1750 lightMeter;
+
+// LiDAR setup
+TFMPlus tfluna;
+HardwareSerial lidarSerial(2); // Using serial port 2
+
+// Display setup
+Adafruit_SSD1306 display(SCREEN_WIDTH, SCREEN_HEIGHT, &Wire, OLED_RESET);
+Adafruit_SSD1306 display_ext(SCREEN_WIDTH, SCREEN_HEIGHT_EXT, &Wire, OLED_RESET);
+U8G2_FOR_ADAFRUIT_GFX u8g2;
+U8G2_FOR_ADAFRUIT_GFX u8g2_ext;
+// ---------------------
