+++ include/lenses.h
@@ -1,0 +1,22 @@
+struct Lens
+{
+  int id;
+  String name;
+  int sensor_reading[7];
+  float distance[7];
+  float apertures[9];
+  bool calibrated;
+};
+
+// Lens and film format definitions
+Lens lenses[] = {
+    {5063, "50/6.3", {0, 0, 0, 0, 0, 0, 0}, {1, 1.2, 1.5, 2, 3, 5, 10}, {0, 6.3, 8, 11, 16, 22, 32, 0, 0}, false},
+    {6563, "65/6.3", {270, 256, 243, 231, 219, 210, 203}, {1, 1.2, 1.5, 2, 3, 5, 10}, {0, 6.3, 8, 11, 16, 22, 32, 0, 0}, true},
+    {7556, "75/5.6", {0, 0, 0, 0, 0, 0, 0}, {1, 1.2, 1.5, 2, 3, 5, 10}, {0, 5.6, 8, 11, 16, 22, 32, 45, 0}, false},
+    {9035, "90/3.5", {0, 0, 0, 0, 0, 0, 0}, {1, 1.2, 1.5, 2, 3, 5, 10}, {3.5, 4, 5.6, 8, 11, 16, 22, 32, 0}, false},
+    {10035, "100/3.5", {0, 0, 0, 0, 0, 0, 0}, {1, 1.2, 1.5, 2, 3, 5, 10}, {3.5, 4, 5.6, 8, 11, 16, 22, 32, 0}, false},
+    {10028, "100/2.8", {0, 0, 0, 0, 0, 0, 0}, {1, 1.2, 1.5, 2, 3, 5, 10}, {2.8, 4, 5.6, 8, 11, 16, 22, 32, 0}, false},
+    {12747, "127/4.7", {0, 0, 0, 0, 0, 0, 0}, {1, 1.2, 1.5, 2, 3, 5, 10}, {4.7, 5.6, 8, 11, 16, 22, 32, 45, 64}, false},
+    {15056, "150/5.6", {0, 0, 0, 0, 0, 0, 0}, {1, 1.2, 1.5, 2, 3, 5, 10}, {5.6, 8, 11, 16, 22, 32, 45, 0, 0}, false},
+    {25005, "250/5.0", {0, 0, 0, 0, 0, 0, 0}, {1, 1.2, 1.5, 2, 3, 5, 10}, {5, 8, 11, 16, 22, 32, 45, 0, 0}, false}  
+};
